Split call tests for pallet-dip-provider into multiple files
