+++ pallets/pallet-dip-provider/src/tests/commit_identity.rs
@@ -16,7 +16,7 @@
 
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
-use frame_support::{assert_noop, assert_ok};
+use frame_support::assert_ok;
 use kilt_support::mock::mock_origin::DoubleOrigin;
 
 use crate::mock::*;
@@ -76,41 +76,4 @@ fn commit_identity_override_same_version_commitment() {
 				Some(expected_identity_commitment)
 			);
 		});
-}
-
-#[test]
-fn delete_identity_commitment_multiple_versions() {
-	ExtBuilder::default()
-		.with_commitments(vec![(DID, 0, u32::MAX), (DID, 1, u32::MAX - 1)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(DipProvider::delete_identity_commitment(
-				DoubleOrigin(ACCOUNT_ID, DID).into(),
-				DID,
-				Some(0),
-			));
-			assert_eq!(
-				crate::pallet::IdentityCommitments::<TestRuntime>::iter_key_prefix(&DID).count(),
-				1
-			);
-			assert_ok!(DipProvider::delete_identity_commitment(
-				DoubleOrigin(ACCOUNT_ID, DID).into(),
-				DID,
-				Some(1),
-			));
-			assert_eq!(
-				crate::pallet::IdentityCommitments::<TestRuntime>::iter_key_prefix(&DID).count(),
-				0
-			);
-		});
-}
-
-#[test]
-fn delete_identity_commitment_not_found() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert_noop!(
-			DipProvider::delete_identity_commitment(DoubleOrigin(ACCOUNT_ID, DID).into(), DID, Some(0),),
-			crate::Error::<TestRuntime>::CommitmentNotFound
-		);
-	});
 }
